+++ packages/components/src/internal/components/entities/EntityMoveConfirmationModal.spec.tsx
@@ -4,7 +4,7 @@ import { PermissionTypes } from '@labkey/api';
 
 import { mountWithAppServerContext, waitForLifecycle } from '../../testHelpers';
 import { ConfirmModal } from '../base/ConfirmModal';
-import { TEST_FOLDER_CONTAINER, TEST_PROJECT_CONTAINER } from '../../../test/data/constants';
+import { TEST_FOLDER_CONTAINER, TEST_PROJECT_CONTAINER } from '../../containerFixtures';
 import { getTestAPIWrapper } from '../../APIWrapper';
 import { getSecurityTestAPIWrapper } from '../security/APIWrapper';
 import { Container } from '../base/models/Container';
