+++ packages/components/src/internal/components/project/DeleteProjectModal.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Button } from 'react-bootstrap';
 
-import { TEST_FOLDER_CONTAINER } from '../../../test/data/constants';
+import { TEST_FOLDER_CONTAINER } from '../../containerFixtures';
 
 import { TEST_USER_APP_ADMIN } from '../../userFixtures';
